Support map rotation #669 - Added traces layer and ability to manipulate.

+++ IsraelHiking.Web/sources/application/services/traces.service.ts
@@ -1,54 +1,29 @@
 import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
-import { Subject } from "rxjs";
-import { remove } from "lodash";
+import { NgRedux } from "@angular-redux/store";
 
-import { DataContainer } from "../models/models";
+import { Trace, ApplicationState } from "../models/models";
 import { Urls } from "../urls";
-
-
-export type Visibility = "private" | "public" | "local";
-
-export interface ITrace {
-    name: string;
-    user: string;
-    description: string;
-    url: string;
-    imageUrl: string;
-    dataUrl: string;
-    id: string;
-    timeStamp: Date;
-    tags: string[];
-    tagsString: string;
-    visibility: Visibility;
-    isInEditMode: boolean;
-}
+import { RemoveTraceAction, UpdateTraceAction, AddTraceAction } from "../reducres/traces.reducer";
 
 @Injectable()
 export class TracesService {
 
-    public readOnlyDataContainer: DataContainer;
-    public missingParts: GeoJSON.FeatureCollection<GeoJSON.LineString>;
-
-    public traces: ITrace[];
-    public tracesChanged: Subject<any>;
-
-    constructor(private readonly httpClient: HttpClient) {
-        this.traces = [];
-        this.tracesChanged = new Subject();
+    constructor(private readonly httpClient: HttpClient,
+        private readonly ngRedux: NgRedux<ApplicationState>) {
     }
 
-    public getMissingParts(trace: ITrace): Promise<GeoJSON.FeatureCollection<GeoJSON.LineString>> {
+    public getMissingParts(trace: Trace): Promise<GeoJSON.FeatureCollection<GeoJSON.LineString>> {
         return this.httpClient.post(Urls.osm + "?url=" + trace.dataUrl, {})
             .toPromise() as Promise<GeoJSON.FeatureCollection<GeoJSON.LineString>>;
     }
 
-    public getTraces = async (): Promise<any> => {
+    public syncTraces = async (): Promise<any> => {
         try {
-            let response = await this.httpClient.get(Urls.osmTrace).toPromise() as ITrace[];
-            this.traces.splice(0);
-            let files = ([] as ITrace[]).concat(response || []);
-            for (let traceJson of files) {
+            let response = await this.httpClient.get(Urls.osmTrace).toPromise() as Trace[];
+            let traces = ([] as Trace[]).concat(response || []);
+            let existingTraces = this.ngRedux.getState().traces.traces;
+            for (let traceJson of traces) {
                 let url = `https://www.openstreetmap.org/user/${traceJson.user}/traces/${traceJson.id}`;
                 let dataUrl = `https://www.openstreetmap.org/api/0.6/gpx/${traceJson.id}/data`;
                 traceJson.url = url;
@@ -57,25 +32,34 @@ export class TracesService {
                 traceJson.dataUrl = dataUrl;
                 traceJson.timeStamp = new Date(traceJson.timeStamp);
                 traceJson.isInEditMode = false;
-                this.traces.push(traceJson);
+                let existingTrace = existingTraces.find(t => t.id === traceJson.id);
+                if (existingTrace != null) {
+                    traceJson.dataContainer = existingTrace.dataContainer;
+                    this.ngRedux.dispatch(new UpdateTraceAction({ traceId: traceJson.id, trace: traceJson }));
+                } else {
+                    this.ngRedux.dispatch(new AddTraceAction({ trace: traceJson }));
+                }
             }
-            this.tracesChanged.next();
+            for (let trace of existingTraces.filter(t => t.visibility !== "local")) {
+                if (traces.find(t => t.id === trace.id) == null) {
+                    this.ngRedux.dispatch(new RemoveTraceAction({ traceId: trace.id }));
+                }
+            }
         } catch (ex) {
             console.error("Unable to get user's traces.");
         }
     }
 
-    public updateTrace = (trace: ITrace): Promise<any> => {
+    public updateTrace = async (trace: Trace): Promise<any> => {
         trace.tags = trace.tagsString.split(",").map(t => t.trim()).filter(t => t);
-        return this.httpClient.put(Urls.osmTrace + trace.id, trace, { responseType: "text" }).toPromise();
+        await this.httpClient.put(Urls.osmTrace + trace.id, trace, { responseType: "text" }).toPromise();
+        this.ngRedux.dispatch(new UpdateTraceAction({ traceId: trace.id, trace: trace }));
     }
 
-    public deleteTrace = (trace: ITrace): Promise<any> => {
-        let promise = this.httpClient.delete(Urls.osmTrace + trace.id, { responseType: "text" }).toPromise();
-        promise.then(() => {
-            remove(this.traces, traceToFind => traceToFind.id === trace.id);
-            this.tracesChanged.next();
-        });
-        return promise;
+    public deleteTrace = async (trace: Trace): Promise<any> => {
+        if (trace.visibility !== "local") {
+            await this.httpClient.delete(Urls.osmTrace + trace.id, { responseType: "text" }).toPromise();
+        }
+        this.ngRedux.dispatch(new RemoveTraceAction({ traceId: trace.id }));
     }
 }
